delete page in health
