+++ components/doctor-now-component/doctor-now-step3.jsx
@@ -5,12 +5,12 @@ import { Box, Button } from "@mui/material";
 import ArrowCircleRightOutlinedIcon from "@mui/icons-material/ArrowCircleRightOutlined";
 import { useTheme } from "@mui/material/styles";
 
-import Screen from "../../public/images/doctor-now-page-3.png";
+import Screen from "../../public/images/doctor-now-page-4.png";
 
 import MissClickPopup from "../miss-click-popup";
 import { glow } from "../glow";
 
-export default function DoctorNowPage3({ handleNext, handlers }) {
+export default function DoctorNowPage4({ handleNext, handlers }) {
   const theme = useTheme();
   const [missClicksCount, setMissclickCount] = useState(0);
   const [openModal, setOpenModal] = useState(false);
@@ -57,9 +57,10 @@ export default function DoctorNowPage3({ handleNext, handlers }) {
           position: "absolute",
           border: `5px solid ${theme.palette.primary.main}`,
           borderRadius: "12px",
-          py: "7%",
-          px: "46%",
-          bottom: "11%",
+          px: "9%",
+          py: "10%",
+          bottom: "47%",
+          left: "10%",
           animation: `${glow} 2s infinite`,
         }}
       ></Box>
