--- components/doctor-now-component/doctor-now-step9.jsx
@@ -1,76 +1,0 @@
-"use client";
-
-import React, { useState } from "react";
-import { Box, Button } from "@mui/material";
-import ArrowCircleRightOutlinedIcon from "@mui/icons-material/ArrowCircleRightOutlined";
-import { useTheme } from "@mui/material/styles";
-
-import Screen from "../../public/images/doctor-now-page-9.png";
-
-import MissClickPopup from "../miss-click-popup";
-import { glow } from "../glow";
-
-export default function DoctorNowPage9({ handleNext, handlers }) {
-  const theme = useTheme();
-  const [missClicksCount, setMissclickCount] = useState(0);
-  const [openModal, setOpenModal] = useState(false);
-
-  const handleMisClick = (event) => {
-    if (
-      !event.target.closest(".clickable-box") &&
-      !event.target.closest(".arrow-button") &&
-      !openModal
-    ) {
-      setMissclickCount((prevCount) => prevCount + 1);
-      if (missClicksCount >= 5) {
-        setOpenModal(true);
-      }
-    }
-  };
-
-  const handleClose = () => {
-    setOpenModal(false);
-    setMissclickCount(0);
-  };
-
-  return (
-    <Box
-      onClick={handleMisClick}
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        backgroundImage: `url(${Screen.src})`,
-        backgroundSize: "contain",
-        backgroundRepeat: "no-repeat",
-        backgroundPosition: "bottom",
-        height: "100vh",
-        width: "97vw",
-        position: "relative",
-      }}
-      {...handlers}
-    >
-      <Box
-        onClick={handleNext}
-        className="clickable-box"
-        sx={{
-          position: "absolute",
-          border: `5px solid ${theme.palette.primary.main}`,
-          borderRadius: "12px",
-          px: "48%",
-          py: "8%",
-          bottom: "0%",
-          animation: `${glow} 2s infinite`,
-        }}
-      ></Box>
-      <Button
-        onClick={handleNext}
-        className="arrow-button"
-        sx={{ position: "absolute", bottom: "1rem", right: "0rem" }}
-      >
-        <ArrowCircleRightOutlinedIcon sx={{ width: "3rem", height: "3rem" }} />
-      </Button>
-      <MissClickPopup handleClose={handleClose} open={openModal} />
-    </Box>
-  );
-}
